TRIB-767 Titan 0.4.0 compatibility and refactoring of tokenizers.
Tokenizers now have two components - a "Setup" component (for now called a GraphBuildingRule) that is used when initializing state for the MR job, an the "MR" component", the actual tokenizer that builds the graph and is dynamically invoked by the mapper.

This was necessitated by the Titan 0.4.0 changes that require graphs to be declared with a "schema" up front before we can add edges and properties.

The tokenizer determines the schema (since it determines the property graph elements) so we had to add code to calculate this schema when creating the MR job. This in turn led to the logical separation of the tokenizer's setup and MRtimee components.

Also, command line validation:
- are column family names specified legal?
- is configuration file specified present?