[TRIB-1145] Several fixes on the new node image
1. This issue is not related to TRIB-1145 but came across this while
working on the new AMI as described in TRIB-1145.

The .localdomain added by aws upon instance bring-up is not liked
by zk or hadoop, so just remove that in preparation for the cluster.

2. Added some delay in configuration for each service start

3. Added workcount running for hadoop sanity check, as well as titan
gods graph load for rexster/titan sanity check. These do not mean no
more manual configuration, just a start-up test.

4. Consolidate the GAODev script to be consistent with GAOQA script.

Signed-off-by: Yi Zou <yi.zou@intel.com>