[TRIB-1200/TRIB-1243] Reorg the per node configuration files
The existing configuration files are quite confusing,
so recreate the file system layout as whatever looks
like in the real AWS node, so we know exactly where and
which files are used. With this, the existing directories
of sshconf, cronlog, etc. can be removed.

Files/directories under aws/deploy are reorganized as below, with
original files/dirs on the left being replaced by the files/dirs on the
right
 IntelAnalytics             --> config/home/hadoop/IntelAnaltyics
 config/IntelAnalytics.conf --> config/etc/nginx/conf.d/IntelAnalytics.conf
 config/bashrc/.bashrc      --> config/home/hadoop/.bashrc
 config/cluster/*           --> conifg/home/hadoop/IntelAnalytics/*
 config/cronlog/*.crontab   --> config/var/spool/cron/hadoop
 config/cronlog/*.logrotate --> config/etc/logrotate.d/IntelAnalytics
 config/sshconf/*           --> config/home/hadoop/.ssh/*

Newly added configs (for ganglia)
 config/etc/ganglia/gmond.conf
 config/etc/ganglia/gmetat.conf
 php-fpm.d/www.conf

All configuration/scripts used for the server components for IntelAnalytics
softare are now organized under the 'config' subdir exactly the same way as
it is on the actual AWS cluster node. The list of confiugration files for
the server side components are:

/home/hadoop/.bashrc
/home/hadoop/IntelAnalytics/bin
/home/hadoop/IntelAnalytics/hadoop/conf
/home/hadoop/IntelAnalytics/hbase/conf
/home/hadoop/IntelAnalytics/pig/conf
/home/hadoop/IntelAnalytics/titan-server/conf
/home/hadoop/IntelAnalytics/elasticsearch/config
/etc/nginx/IntelAnalytics.cer
/etc/nginx/IntelAnalytics.key
/etc/nginx/conf.d/IntelAnalytics.conf
/etc/php-fpm.d/www.conf
/etc/ganglia/gmond.conf
/etc/ganglia/gmetad.conf
/etc/logrotate.d/IntelAnalytics

Note, client side configuration are part of IntelAnalytics
RPMs.

Signed-off-by: Yi Zou <yi.zou@intel.com>