[TRIB-1275] Fix the ipython socket reconnecting issue in AWS
We have observed that ports used by ipython notebook are being closed and reconnected
every roughly one minute once the not3ebook session starts. This is being root-caused
due to missing several additional configuration to nginx, particularly extending the
default 60s of proxy_read_timeout and proxy_send_timeout, which are generally used with
a large value when transfering large sized files over proxy. We will set the default
timeout to be 1 day, which should be realisticly more enough, and recreating the same
sockets will still happen when timeout happens, so no impact on whatever is being
executed in the ipython kernel.

Note: you need to manually integrate this to your existing cluster's nginx config
file and restart the service under sudo to have this be in effect:

1. ssh as ec2-user to master node
2. sudo vim /etc/nginx/conf.d/IntelAnalytics.conf
3. Add the lines in this patch
4. sudo service nginx reload

Signed-off-by: Yi Zou <yi.zou@intel.com>