[TRIB-1405] Bug fix to make INTEL_ANALYTICS_PYTHON consistent
INTEL_ANALYTICS_PYTHON is inconsistency between rpm install and source code execution.

The /home/hadoop/IntelAnalytics/bin/IntelAnaltyics_env.sh is still using the old virtpy
path where IntelAnalytics used to reside in, fix it to be consistent w/ the current rpm
layout.

- Try to figure out the INTEL_ANALYTICS_HOME from rpm installation query, and set it only
if it does not exist, which we can't help much anyway.
- Update INTEL_ANALYTICS_PYTHON to poin the final settled path
- Remove INTEL_ANALYTICS_JAR, which is set by intel_analytics.properties

This is fixed in the latest Build.12 AWS image as well.

Signed-off-by: Yi Zou <yi.zou@intel.com>