[TRIB-1631] Bug fix on titan rexster server stopping
The configuration file on rexster shutdown-host field is incorrect, so enforce it to be
0.0.0.0 upon cluster configuration time.

For any instances on Build.13 or earlier, please follow the steps below to fix the configuraiton
manaully.

1) login to master node as user 'hadoop'
2) stop the rexster instance, e.g., ~/IntelAnalytics/titan/bin/stop-rexstitan.sh
3) if 2) fails, just do 'jps -ls' and locate the PID of the rexster titan server, and kill -9 the PID directly
4) edit ~/IntelAnalytics/titan/conf/rexstitan-hbase-es.xml, locate the field <shutdown-host>..</shutdown-host>,
and update it to be 127.0.0.1
5) restart the rexster, ~/IntelAnalytics/titan/bin/start-rexstitan.sh

This fix will be rolled into the next AMI build version, Build.14.

Note: this is cherry picked from origin/sprint_7 as it was not pulled in earlier.

Signed-off-by: Yi Zou <yi.zou@intel.com>