[TRIB-1711] Do no enable ganglia services
We are not using ganglia, so stop feeding stats. This only requires
changes in cluster configuration script, and can be easily commented
out if later we want to reenable this.

This fix is expected to drop the memory footprint for both hadoop and
hbase caused the py periodic stats feeding to ganglia.

Signed-off-by: Yi Zou <yi.zou@intel.com>