[TRIB-562] PIG-3276 patch cherry-picked and ported for Pig 0.12.0 source
This is the patch cherry-picked from Pig upstream source and ported to
apply cleanly on Pig 0.12.0 source tree. We need this for 0.8 release.
This patch fixes a Null pointer exception from Pig caused by a bug in
RANK on an empty set.

More details can be found at:

https://jira01.devtools.intel.com/browse/TRIB-562
https://issues.apache.org/jira/browse/PIG-3726

This commit contains the patch file that can be used to directly apply
to the 0.12.0 pig source tree by:

> cd ${pig-0.12.0-source}
> patch -p1 < ${path-to-patch-folder}/PIG-3276-cherry-picked-for-tribeca-0.8.
> ant

Changes to be committed:
	new file:   aws/deploy/patches/PIG-3276-cherry-picked-for-tribeca-0.8