[TRIB-562] Use the PigStatesUtil instead of MRPigStatsUtil
This is the fix needed to backport the PIG-3726 upstream patch to get
it to build for 0.12.0 Pig.

The build would fail since Pig 0.12.0 uses PigStatsUtil as there is no
MRPigStatsUtil, so replace MRPigStatsUtil w/ the PigStatsUitl.

After this, generate the 0.12.0 build like below, for example:

> ant -Dpig.version=0.12.0 -Dpig.version.suffix="-JIRA-3256"

to produce the updated jars.
Signed-off-by: Yi Zou <yi.zou@intel.com>