release_0.8.0_die_apl_die
Removal of average path length function from 0.8.0:  Under consutlation of Kathleen and Bryn, we agreed that since the giraph code only provides a first step that the user then has to augment with a yet unprovided Gremlin script, APL would not be released in 0.8.0.  We provide full functionality; not partial work that the user has to finish off with a script in a different tool.

Cleaned up comments in numerical statistics section.