make changes to the poms for engine-spark, graphbuilder-3, and graphon to eliminate a jar dependenciy issue.
Spark-1.0 utilizes Servlet 3.0 libraries which conflict with servlet2-5 libraries which are commonly found in other dependencies. This does not affect the compiled application because we set the classpath before executing the application but the classpath for unit tests is set by maven. This causes several unit test suites to abort because of a security exception caused by the class javax.servlet.FilterRegistration to have a different jar signature to other classes in the javax.servlet package. This is resolved by excluding the artifacts org.mortbay.jetty:servlet-api-2.5 and javax.servlet:servlet-api in the three maven projects