TRIB-3475
Brought back the state space size argument.

It is now used to validate the incoming data: It is an exception for a property to have its prior vector of length different from the specified state space.

If the property containing the prior is not on a vertex, that causes a NotFoundException.