TRIB-3876
Replaced the "clique to clique" unipartite graph construction with a bipartite "clique to shadow" constuction. This *should* greatly reduce the sizes of graphs that must be constructed when the typical (k-1) clique has many extensions to a k-clique (in the most extreme cases this saves a factor of the vertex count... 1000s or millions or even... BEEEELYUNS).

We'll yet see if it yields the hoped for performance gains.

This forced some renamings and some rewriting of unit tests.